Rename `Reader` and `Writer` (BREAKING CHANGE)
These renamings are as per the discussion at
https://github.com/BurntSushi/rust-snappy/issues/17, and they're
intended to match the API of the `flate2` crate.
