+++ examples/compress.rs
@@ -8,6 +8,6 @@ fn main() {
 
     let mut rdr = stdin.lock();
     // Wrap the stdout writer in a Snappy writer.
-    let mut wtr = snap::Writer::new(stdout.lock());
+    let mut wtr = snap::write::FrameEncoder::new(stdout.lock());
     io::copy(&mut rdr, &mut wtr).expect("I/O operation failed");
 }
