+++ src/tests.rs
@@ -386,19 +386,19 @@ fn depress(bytes: &[u8]) -> Vec<u8> {
 
 fn frame_press(bytes: &[u8]) -> Vec<u8> {
     use std::io::Write;
-    use frame::Writer;
+    use write;
 
-    let mut wtr = Writer::new(vec![]);
+    let mut wtr = write::FrameEncoder::new(vec![]);
     wtr.write_all(bytes).unwrap();
     wtr.into_inner().unwrap()
 }
 
 fn frame_depress(bytes: &[u8]) -> Vec<u8> {
     use std::io::Read;
-    use frame::Reader;
+    use read;
 
     let mut buf = vec![];
-    Reader::new(bytes).read_to_end(&mut buf).unwrap();
+    read::FrameDecoder::new(bytes).read_to_end(&mut buf).unwrap();
     buf
 }
